chore: renamed dashboard-components to dashboard
