+++ apps/web/app/(authenticated)/(app)/app/layout.tsx
@@ -1,4 +1,4 @@
-import { DesktopSidebar } from "@/components/dashboard-components/desktop-sidebar";
+import { DesktopSidebar } from "@/components/dashboard/desktop-sidebar";
 import { getTenantId } from "@/lib/auth";
 import { db, eq, schema } from "@unkey/db";
 import { redirect } from "next/navigation";
