+++ apps/web/app/(authenticated)/(app)/layout.tsx
@@ -1,5 +1,5 @@
 import { Toaster } from "@/components/ui/toaster";
-import { ReactQueryProvider } from "@/components/dashboard-components/react-query-provider";
+import { ReactQueryProvider } from "@/components/dashboard/react-query-provider";
 
 export default function Layout({ children }: { children: React.ReactNode }) {
   return (
